backend/python_services/embedding_service/requirements.txt
Flask>=3.0.0
sentence-transformers>=2.2.2
torch>=2.0.0
accelerate>=0.21.0
python-dotenv>=1.0.1
gunicorn>=21.2.0
firebase-admin>=6.5.0